Placeholder stuff in builder script.
Commented out the code that will in the future be used to increment the
version number using NPM's native "version" command. This will update
package.json and hopefully callback with the new semver string which we can
use the same way as before when creating the new directories and replacing
the magic strings in various files. Manually set the version in package.json
to 0.1.24 which is the current version on literal-trees.co.