Calculate viewBox in separate module
tree-sizing.js will calculate the appropriate view box, taking into account
tree.options.aspectratio if it is defined. If aspectratio is undefined the
function will try to guess if this is a small mobile device where it makes
the most sense to return a view box which matches the aspect ratio of the
device. On desktop browsers with large screens it isn't as important.

Usage:

    var TreeSizing = require('./tree-sizing');

    var viewBox = TreeSizing.calculateViewBox({
        aspectratio: tree.options.aspectratio,
        treeWidth: treeWidth,
        treeHeight: treeHeight,
        margin: tree.options.margin,
        treeMinX: tree.TREE_MIN_X,
        treeMinY: tree.TREE_MIN_Y,
        sizeAdjustmentMethod: tree.options.sizeAdjustmentMethod
    });