Adjust size to "contain"
The tree-sizing module checks if the screen aspect ratio is tall and skinny.
If so, it tried to re-position the x, y, width, and height of the viewBox to
make the box match the aspect ratio of the screen.

Querystring param usage is ?sizingmethod=contain. No support for
sizingmethod=cover yet but hope to soon.