Fix very weird flicker of btf content
For some reason using the nyt-cheltenham font is causing a layout flicker in
the Inputs table below the fold. Not sure why since the font face data is
inlined as a base64 data URL.

Solving the problem by making the #btf div visibility:hidden and then
setting it to "visible" in an inline script tag immediately below the
content. There is a 100ms CSS fade in effect just for the hell of it.