Finally get "new tree" reload logic correct (?)
Well, hopefully. Probably making this way harder than it needs to be. Does a
string comparison between the old query string and the new. If they are the
same it calls window.location.reload(), if they are different it just sets
window.location.search and lets the browser perform the reload.